+++ RNSentryTester/RNSentryTesterTests/RNSentry+initNativeSdk.mm
@@ -1,6 +1,7 @@
 #import <UIKit/UIKit.h>
 #import <XCTest/XCTest.h>
 #import <Sentry/SentryOptions.h>
+#import <Sentry/SentryEvent.h>
 #import "RNSentry.h"
 
 @interface RNSentryInitNativeSdkTests : XCTestCase
@@ -17,8 +18,11 @@
     NSDictionary *_Nonnull mockedReactNativeDictionary = @{
         @"dsn": @"https://abcd@efgh.ingest.sentry.io/123456",
         @"beforeSend": @"will_be_overwritten",
-        @"enableNativeCrashHandling": @YES,
-
+        @"tracesSampleRate": @1,
+        @"tracesSampler": ^(SentrySamplingContext *_Nonnull samplingContext) {
+          return @1;
+        },
+        @"enableTracing": @YES,
     };
     SentryOptions* actualOptions = [rnSentry createOptionsWithDictionary:mockedReactNativeDictionary error:&error];
 
@@ -27,6 +31,7 @@
     XCTAssertNotNil(actualOptions.beforeSend, @"Before send is overwriten by the native RNSentry implementation");
     XCTAssertEqual(actualOptions.tracesSampleRate, nil, @"Traces sample rate should not be passed to native");
     XCTAssertEqual(actualOptions.tracesSampler, nil, @"Traces sampler should not be passed to native");
+    XCTAssertEqual(actualOptions.enableTracing, false, @"EnableTracing should not be passed to native");
 }
 
 - (void)testCreateOptionsWithDictionaryNativeCrashHandlingDefault
@@ -43,7 +48,7 @@
     XCTAssertEqual([actualOptions.integrations containsObject:@"SentryCrashIntegration"], true, @"Did not set native crash handling");
 }
 
-- (void)testCreateOptionsWithDictionaryPerformanceTrackingDefault
+- (void)testCreateOptionsWithDictionaryAutoPerformanceTracingDefault
 {
     RNSentry * rnSentry = [[RNSentry alloc] init];
     NSError* error = nil;
@@ -72,7 +77,7 @@
     XCTAssertEqual([actualOptions.integrations containsObject:@"SentryCrashIntegration"], true, @"Did not set native crash handling");
 }
 
-- (void)testCreateOptionsWithDictionaryPerformanceTrackingEnabled
+- (void)testCreateOptionsWithDictionaryAutoPerformanceTracingEnabled
 {
     RNSentry * rnSentry = [[RNSentry alloc] init];
     NSError* error = nil;
@@ -80,7 +85,6 @@
     NSDictionary *_Nonnull mockedReactNativeDictionary = @{
         @"dsn": @"https://abcd@efgh.ingest.sentry.io/123456",
         @"enableAutoPerformanceTracing": @YES,
-
     };
     SentryOptions* actualOptions = [rnSentry createOptionsWithDictionary:mockedReactNativeDictionary error:&error];
     XCTAssertNotNil(actualOptions, @"Did not create sentry options");
@@ -96,7 +100,6 @@
     NSDictionary *_Nonnull mockedReactNativeDictionary = @{
         @"dsn": @"https://abcd@efgh.ingest.sentry.io/123456",
         @"enableNativeCrashHandling": @NO,
-
     };
     SentryOptions* actualOptions = [rnSentry createOptionsWithDictionary:mockedReactNativeDictionary error:&error];
     XCTAssertNotNil(actualOptions, @"Did not create sentry options");
@@ -104,7 +107,7 @@
     XCTAssertEqual([actualOptions.integrations containsObject:@"SentryCrashIntegration"], false, @"Did not disable native crash handling");
 }
 
-- (void)testCreateOptionsWithDictionaryPerformanceTrackingDisabled
+- (void)testCreateOptionsWithDictionaryAutoPerformanceTracingDisabled
 {
     RNSentry * rnSentry = [[RNSentry alloc] init];
     NSError* error = nil;
@@ -112,7 +115,6 @@
     NSDictionary *_Nonnull mockedReactNativeDictionary = @{
         @"dsn": @"https://abcd@efgh.ingest.sentry.io/123456",
         @"enableAutoPerformanceTracing": @NO,
-
     };
     SentryOptions* actualOptions = [rnSentry createOptionsWithDictionary:mockedReactNativeDictionary error:&error];
     XCTAssertNotNil(actualOptions, @"Did not create sentry options");
@@ -126,12 +128,44 @@
     NSError* error = nil;
 
     NSDictionary *_Nonnull mockedReactNativeDictionary = @{
-        @"dsn": @"not_a_valid_dsn"
+        @"dsn": @"not_a_valid_dsn",
     };
     SentryOptions* actualOptions = [rnSentry createOptionsWithDictionary:mockedReactNativeDictionary error:&error];
 
     XCTAssertNil(actualOptions, @"Created invalid sentry options");
     XCTAssertNotNil(error, @"Did not created error on invalid dsn");
+}
+
+- (void)testEventFromSentryCocoaReactNativeHasOriginAndEnvironmentTags
+{
+  RNSentry* rnSentry = [[RNSentry alloc] init];
+  SentryEvent* testEvent = [[SentryEvent alloc] init];
+  testEvent.sdk = @{
+    @"name": @"sentry.cocoa.react-native",
+  };
+
+  [rnSentry setEventOriginTag: testEvent];
+  
+  XCTAssertEqual(testEvent.tags[@"event.origin"], @"ios");
+  XCTAssertEqual(testEvent.tags[@"event.environment"], @"native");
+}
+
+- (void)testEventFromSentryReactNativeOriginAndEnvironmentTagsAreOverwritten
+{
+  RNSentry* rnSentry = [[RNSentry alloc] init];
+  SentryEvent* testEvent = [[SentryEvent alloc] init];
+  testEvent.sdk = @{
+    @"name": @"sentry.cocoa.react-native",
+  };
+  testEvent.tags = @{
+    @"event.origin": @"testEventOriginTag",
+    @"event.environment": @"testEventEnvironmentTag",
+  };
+  
+  [rnSentry setEventOriginTag: testEvent];
+  
+  XCTAssertEqual(testEvent.tags[@"event.origin"], @"ios");
+  XCTAssertEqual(testEvent.tags[@"event.environment"], @"native");
 }
 
 @end