+++ android/src/main/java/io/sentry/react/RNSentryModuleImpl.java
@@ -10,7 +10,6 @@ import android.content.res.AssetManager;
 import android.util.SparseIntArray;
 
 import androidx.annotation.Nullable;
-import androidx.annotation.NonNull;
 import androidx.core.app.FrameMetricsAggregator;
 
 import com.facebook.react.bridge.Arguments;
@@ -75,6 +74,8 @@ public class RNSentryModuleImpl {
 
     private static JavaScriptExecutorFactory javaScriptExecutorFactory = null;
 
+    private static final String NATIVE_SDK_NAME = "sentry.native.android";
+    private static final String ANDROID_SDK_NAME = "sentry.java.android.react-native";
     private static final ILogger logger = new AndroidLogger(NAME);
     private static final BuildInfoProvider buildInfo = new BuildInfoProvider(logger);
     private static final String modulesPath = "modules.json";
@@ -114,12 +115,11 @@ public class RNSentryModuleImpl {
 
     public void initNativeSdk(final ReadableMap rnOptions, Promise promise) {
         SentryAndroid.init(this.getReactApplicationContext(), options -> {
-            @NonNull final String sdkName = "sentry.java.android.react-native";
             @Nullable SdkVersion sdkVersion = options.getSdkVersion();
             if (sdkVersion == null) {
-                sdkVersion = new SdkVersion(sdkName, BuildConfig.VERSION_NAME);
+                sdkVersion = new SdkVersion(ANDROID_SDK_NAME, BuildConfig.VERSION_NAME);
             } else {
-                sdkVersion.setName(sdkName);
+                sdkVersion.setName(ANDROID_SDK_NAME);
             }
 
             options.setSentryClientName(sdkVersion.getName() + "/" + sdkVersion.getVersion());
@@ -671,10 +671,10 @@ public class RNSentryModuleImpl {
             switch (sdk.getName()) {
                 // If the event is from capacitor js, it gets set there and we do not handle it
                 // here.
-                case "sentry.native":
+                case NATIVE_SDK_NAME:
                     setEventEnvironmentTag(event, "native");
                     break;
-                case "sentry.java.android":
+                case ANDROID_SDK_NAME:
                     setEventEnvironmentTag(event, "java");
                     break;
                 default: